Change Test, remove empty result and add check size to return

+++ src/main/java/com/aleksandrbogomolov/controller/ContactController.java
@@ -33,12 +33,12 @@ public class ContactController {
     @RequestMapping(method = RequestMethod.GET)
     public ResponseEntity<List<Contact>> getFilteredContact(
             @RequestParam(name = "nameFilter") String regex,
-            @RequestParam(name = "lastId", required = false) Integer lastId,
+            @RequestParam(name = "lastId", required = false) Long lastId,
             @RequestParam(name = "limit", required = false) Integer limit) {
 
         log.info("Get filtered contacts");
 
-        int lastIdParam = lastId == null ? 0 : lastId;
+        long lastIdParam = lastId == null ? 0 : lastId;
         int limitParam = limit == null || limit > 5 ? 5 : limit;
 
         List<Contact> contacts = service.getFilteredContacts(regex, lastIdParam, limitParam);
