+++ src/main/java/com/aleksandrbogomolov/repository/jdbc/ContactRepositoryImpl.java
@@ -28,7 +28,7 @@ public class ContactRepositoryImpl implements ContactRepository {
     }
 
     @Override
-    public List<Contact> getLimitAll(int lastId, int limit) {
+    public List<Contact> getLimitAll(long lastId, int limit) {
         return template.query("SELECT * FROM contacts WHERE id > ? ORDER BY id LIMIT ?", mapper, lastId, limit);
     }
 }
